refactor(web): unify auth surface — AuthModal becomes the glassy overlay
The auth modal was inconsistent with the new sign-up overlay, so guests
were seeing two different surfaces depending on where they triggered
sign-in (project gate, settings gate, sidebar Sign up button used the
old card; the chat input limit used the new overlay). Folded the new
look into AuthModal so every entry point now opens the same full-page
glassy presentation: GuestGate gates, sidebar Sign up, MainContent
guest-limit, MainContent session expired, and the ImageGalleryView
sign-in prompt all pick it up automatically.

Mechanics: rewrote AuthModal's overlay/modal chrome to the page-level
glassy treatment with a Back affordance in place of the close button.
All form features preserved — Google OAuth, sign-in/up tabs (now a
subtle footer toggle), email + password, name (sign-up), Remember me
+ Forgot password row (sign-in), forgot-password reset flow, password
show/hide, error/success states, Esc and backdrop dismiss. Kept
.loadingOverlay/.loadingLogo/.loadingSpinner classes since App.jsx
imports this stylesheet for its bootstrap loader.

GuestLimitOverlay was a transitional duplicate; deleted. MainContent
guest-limit usage points back at AuthModal with initialTab="signup".
