--- src/components/GuestLimitOverlay/GuestLimitOverlay.module.css
@@ -1,347 +1,0 @@
-/* GuestLimitOverlay — full-screen, glassy sign-up surface.
-   Sits over the current page with a heavy, theme-aware blur so the
-   underlying chat fades into the background without being obscured. */
-
-.overlay {
-  position: fixed;
-  inset: 0;
-  z-index: 2000;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 32px 24px calc(32px + var(--safe-area-bottom));
-  background-color: rgba(253, 249, 243, 0.62);
-  backdrop-filter: blur(28px) saturate(140%);
-  -webkit-backdrop-filter: blur(28px) saturate(140%);
-  animation: overlayIn 0.2s ease;
-}
-
-:global([data-theme='dark']) .overlay {
-  background-color: rgba(26, 26, 26, 0.62);
-}
-
-@keyframes overlayIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-/* Back button — top-left dismiss affordance. */
-.backBtn {
-  position: absolute;
-  top: calc(20px + var(--safe-area-top));
-  left: calc(20px + var(--safe-area-left));
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  padding: 8px 14px 8px 12px;
-  border-radius: 10px;
-  background-color: var(--bg-button);
-  border: 1px solid var(--border-color);
-  color: var(--text-secondary);
-  font-size: 13px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: background-color 0.15s ease, color 0.15s ease, transform 0.1s ease;
-}
-
-.backBtn:hover {
-  background-color: var(--bg-button-hover);
-  color: var(--text-primary);
-}
-
-.backBtn:active {
-  transform: scale(0.97);
-}
-
-.backBtn:focus-visible {
-  outline: none;
-  border-color: var(--text-secondary);
-}
-
-/* Panel — centered content; the surface sits on its own (no card chrome)
-   so it reads as part of the page rather than a stacked modal. */
-.panel {
-  width: 100%;
-  max-width: 380px;
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  text-align: center;
-  animation: panelIn 0.3s cubic-bezier(0.2, 0, 0, 1);
-}
-
-@keyframes panelIn {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.logoMark {
-  width: 48px;
-  height: 48px;
-  border-radius: 14px;
-  background-color: var(--bg-icon-button);
-  color: var(--text-icon-button);
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 22px;
-  font-weight: 700;
-  margin: 0 auto 18px;
-}
-
-.heading {
-  font-size: 24px;
-  font-weight: 700;
-  letter-spacing: -0.01em;
-  line-height: 1.25;
-  color: var(--text-primary);
-  margin: 0 0 8px;
-}
-
-.subtitle {
-  font-size: 14px;
-  line-height: 1.55;
-  color: var(--text-secondary);
-  margin: 0 0 24px;
-}
-
-/* Continue with Google */
-.googleBtn {
-  width: 100%;
-  padding: 12px 16px;
-  border-radius: 12px;
-  border: 1px solid var(--border-color);
-  background-color: var(--bg-button);
-  cursor: pointer;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 10px;
-  font-size: 14px;
-  font-weight: 600;
-  color: var(--text-primary);
-  transition: background-color 0.15s ease, border-color 0.15s ease, transform 0.1s ease;
-}
-
-.googleBtn:hover {
-  background-color: var(--bg-button-hover);
-  border-color: var(--border-input);
-}
-
-.googleBtn:active {
-  transform: scale(0.99);
-}
-
-.googleBtn:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-  transform: none;
-}
-
-.googleIcon {
-  width: 18px;
-  height: 18px;
-  flex-shrink: 0;
-}
-
-/* Divider */
-.divider {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  margin: 18px 0;
-}
-
-.dividerLine {
-  flex: 1;
-  height: 1px;
-  background-color: var(--border-color);
-}
-
-.dividerText {
-  font-size: 12px;
-  font-weight: 500;
-  color: var(--text-muted);
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-}
-
-/* Form */
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-  text-align: left;
-}
-
-.fieldGroup {
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-}
-
-.label {
-  font-size: 13px;
-  font-weight: 600;
-  color: var(--text-secondary);
-}
-
-.input {
-  width: 100%;
-  padding: 11px 14px;
-  border-radius: 10px;
-  border: 1px solid var(--border-input);
-  background-color: var(--bg-input);
-  color: var(--text-primary);
-  font-size: 14px;
-  outline: none;
-  transition: border-color 0.15s ease, box-shadow 0.15s ease;
-  box-sizing: border-box;
-}
-
-.input::placeholder {
-  color: var(--text-muted);
-}
-
-.input:focus {
-  border-color: var(--text-secondary);
-  box-shadow: 0 0 0 3px rgba(0, 0, 0, 0.04);
-}
-
-:global([data-theme='dark']) .input:focus {
-  box-shadow: 0 0 0 3px rgba(255, 255, 255, 0.06);
-}
-
-.passwordWrapper {
-  position: relative;
-}
-
-.passwordWrapper .input {
-  padding-right: 40px;
-}
-
-.passwordToggle {
-  position: absolute;
-  right: 10px;
-  top: 50%;
-  transform: translateY(-50%);
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 4px;
-  color: var(--text-muted);
-  display: flex;
-  align-items: center;
-  transition: color 0.15s ease;
-}
-
-.passwordToggle:hover {
-  color: var(--text-secondary);
-}
-
-.submitBtn {
-  width: 100%;
-  padding: 12px 16px;
-  border-radius: 12px;
-  border: none;
-  background-color: var(--bg-icon-button);
-  color: var(--text-icon-button);
-  font-size: 14px;
-  font-weight: 600;
-  cursor: pointer;
-  transition: opacity 0.15s ease, transform 0.1s ease;
-  margin-top: 4px;
-}
-
-.submitBtn:hover {
-  opacity: 0.9;
-}
-
-.submitBtn:active {
-  transform: scale(0.99);
-}
-
-.submitBtn:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-  transform: none;
-}
-
-.spinner {
-  display: inline-block;
-  width: 16px;
-  height: 16px;
-  border: 2px solid transparent;
-  border-top-color: currentColor;
-  border-radius: 50%;
-  animation: spin 0.6s linear infinite;
-}
-
-@keyframes spin {
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-/* Status messages */
-.error,
-.success {
-  font-size: 13px;
-  text-align: center;
-  padding: 8px 12px;
-  border-radius: 8px;
-}
-
-.error {
-  color: #ef4444;
-  background-color: rgba(239, 68, 68, 0.08);
-}
-
-.success {
-  color: #22c55e;
-  background-color: rgba(34, 197, 94, 0.08);
-}
-
-/* Footer link to toggle sign-in / sign-up. */
-.footer {
-  margin: 18px 0 0;
-  font-size: 13px;
-  color: var(--text-muted);
-  text-align: center;
-}
-
-.footerLink {
-  background: none;
-  border: none;
-  padding: 0;
-  cursor: pointer;
-  color: var(--text-primary);
-  font-weight: 600;
-  font-size: 13px;
-  text-decoration: underline;
-  text-underline-offset: 2px;
-}
-
-.footerLink:hover {
-  opacity: 0.8;
-}
-
-/* Compact spacing on small viewports */
-@media (max-width: 480px) {
-  .overlay {
-    padding: 24px 20px calc(24px + var(--safe-area-bottom));
-  }
-  .heading {
-    font-size: 22px;
-  }
-}
