--- src/components/GuestLimitOverlay/index.js
@@ -1,2 +1,0 @@
-export { default as GuestLimitOverlay } from './GuestLimitOverlay';
-export { default } from './GuestLimitOverlay';
